deepen: inline welcome-service into onboarding action
Replaces the one-caller `lib/email/welcome-service.ts` module
(exporting `queueWelcomeEmail`) with a private helper function of
the same name inside its sole caller,
`app/(auth)/onboarding/actions.ts:completeOnboardingBatched`.

Why this is a deepening
-----------------------
`queueWelcomeEmail` was extract-for-testability with exactly one
production caller. Its interface (userId, email, name -> void promise)
was nearly as complex as the body, which itself is a fire-and-forget
`setImmediate` wrapper around one `prisma.user.findUnique`, one
`getEmailTemplate(EmailType.WELCOME, ...)`, and one `sendEmail(...)`.

Applies the deletion test: folding the body into its caller
concentrates the logic in one place; nothing reappears across N
callers because there are no N callers. Same shape as ADR-0002
(Inline Analysis-Depth Scoring) and the CONTEXT.md
"Why It Matters Guard" / "Ticker Grounding Guard" / "Historical Context"
precedents. Follows the "-service.ts wraps a Prisma+email chain with 1
caller" pattern the nightly routine has been consolidating away.

Dependency category
-------------------
In-process (Prisma + Resend, both already reached from
`onboarding/actions.ts` for the sibling `sendOnboardingFirstEmail`
call). No adapter change. No new imports leak out of the module's
former boundary — they land in the file whose action already owns
the onboarding write path.

Tests
-----
- `__tests__/email/reply-to-regression.test.ts` retargeted at the new
  location. The source-grep style is preserved (matches
  `replyTo: FOUNDER_REPLY_TO | 'wilf@tldrsec.app'` in the
  `queueWelcomeEmail` body). Four cases pass. No test coverage lost.
- `lib/email/config.ts` docstring pointer updated to reference
  the new inline location.
- `lib/email/welcome-service.ts` deleted.
